Viewの修正
