+++ TikokulympicBeta/Views/EventList/EventEditView.swift
@@ -14,10 +14,7 @@ struct EventEditView: View {
     @Environment(\.presentationMode) var presentationMode
     @State private var isShowingCancelAlert = false
     @State private var isShowingCreateAlert = false
-    @State var autocompleteResults: [GMSAutocompletePrediction]=[]
-    @State var searchtext: String = ""
-    
-    
+
     var body: some View {
         NavigationView {
             VStack {
@@ -53,14 +50,29 @@ struct EventEditView: View {
                             }
                             
                             ContactInfoSection(message: $viewModel.contactInfo)
-                            LocationSearchSection(searchText: $searchtext, autocompleteResults: $autocompleteResults)
-                            ForEach(autocompleteResults, id: \.placeID) { result in
-                            Text(result.attributedPrimaryText.string)
-                                    .onTapGesture {
-                                        viewModel.fetchPlaceDetails(placeID: result.placeID)
-                                        searchtext=result.attributedPrimaryText.string
-                                        autocompleteResults.removeAll()
+                            
+                            VStack(alignment: .leading) {
+                                Text("場所名:")
+                                    .font(.headline)
+                                TextField("場所を入力してください", text: $viewModel.location, onEditingChanged: { isEditing in
+                                    if !isEditing {
+                                        viewModel.autocompleteResults.removeAll()
                                     }
+                                })
+                                .textFieldStyle(RoundedBorderTextFieldStyle())
+                                .padding()
+                                .onChange(of: viewModel.location) { newValue in
+                                    viewModel.fetchAutocompleteResults(input: newValue)
+                                }
+
+                                ForEach(viewModel.autocompleteResults, id: \.placeID) { result in
+                                    Text(result.attributedPrimaryText.string)
+                                        .onTapGesture {
+                                            viewModel.fetchPlaceDetails(placeID: result.placeID)
+                                            viewModel.location = result.attributedPrimaryText.string
+                                            viewModel.autocompleteResults.removeAll()
+                                        }
+                                }
                             }
                         }
                         .padding()
