+++ TikokulympicBeta/Views/EventList/EventEditViewModel.swift
@@ -17,9 +17,11 @@ class EventEditViewModel: ObservableObject {
     @Published var location = ""
     @Published var fee = ""
     @Published var contactInfo = ""
-    @Published var locationSearchQuery = ""
     @Published var latitude  = 0.0
     @Published var longitude = 0.0
+    @Published var searchText = ""
+    @Published var autocompleteResults: [GMSAutocompletePrediction] = []
+    
     let options: [Option] = []
     let placesClient = GMSPlacesClient.shared()
  
@@ -33,21 +35,43 @@ class EventEditViewModel: ObservableObject {
     }
     
     func fetchPlaceDetails(placeID: String) {
-            placesClient.lookUpPlaceID(placeID) { (place, error) in
-                if let error = error {
-                    print("エラー: (error.localizedDescription)")
-                    return
-                }
-
-                if let place = place {
-                    self.latitude = place.coordinate.latitude
-                    self.longitude = place.coordinate.longitude
-                    
-                    print("緯度: \(self.latitude), 経度: \(self.longitude)")
-                }
+        placesClient.lookUpPlaceID(placeID) { (place, error) in
+            if let error = error {
+                print("エラー: \(error.localizedDescription)")
+                return
             }
+
+            if let place = place {
+                self.latitude = place.coordinate.latitude
+                self.longitude = place.coordinate.longitude
+                
+                print("緯度: \(self.latitude), 経度: \(self.longitude)")
+            }
+        }
     }
 
+    func fetchAutocompleteResults(input: String) {
+        guard !input.isEmpty else {
+            self.autocompleteResults.removeAll()
+            return
+        }
+
+        let filter = GMSAutocompleteFilter()
+        filter.type = .address
+        filter.country = "JP"
+
+        placesClient.findAutocompletePredictions(fromQuery: input, filter: filter, sessionToken: nil) { results, error in
+            if let error = error {
+                print("エラー: \(error.localizedDescription)")
+                return
+            }
+
+            DispatchQueue.main.async {
+                print((results as! [GMSAutocompletePrediction]).map{$0.attributedPrimaryText.string} )
+                self.autocompleteResults = results ?? []
+            }
+        }
+    }
 
     func completeEditing() async {
         guard let cost = Int(fee) else {
